requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
azure-core==1.30.1
azure-functions==1.18.0
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dataclasses-json==0.6.4
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
isodate==0.6.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.23
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.13.3
orjson==3.9.15
packaging==23.2
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pypdf==4.1.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4

